changes for better deployment options

+++ fredtm-core/src/main/java/com/fredtm/core/decorator/TimeMeasure.java
@@ -43,7 +43,7 @@ public enum TimeMeasure {
 		return this.value;
 	}
 	
-	public int getFromMillisConverterFactor() {
+	public double getFromMillisConverterFactor() {
 		return this.fromMillisConverterFactor;
 	}
